Move extensionPagePlatform
