+++ src/tsconfig.strictNullChecks.json
@@ -33,8 +33,6 @@
     "./auth/authSelectors.ts",
     "./auth/authSlice.ts",
     "./auth/authTypes.ts",
-    "./background/setToolbarIconFromTheme.ts",
-    "./background/setToolbarIconFromTheme.test.ts",
     "./background/auth/authStorage.ts",
     "./background/auth/codeGrantFlow.ts",
     "./background/auth/implicitGrantFlow.ts",
@@ -49,12 +47,16 @@
     "./background/messenger/strict/api.ts",
     "./background/messenger/strict/registration.ts",
     "./background/setToolbarBadge.test.ts",
+    "./background/setToolbarIconFromTheme.test.ts",
+    "./background/setToolbarIconFromTheme.ts",
     "./background/sidePanel.ts",
     "./background/toolbarBadge.ts",
     "./background/walkthroughModalTrigger.ts",
     "./bricks/available.ts",
     "./bricks/effects/AddTextCommand.ts",
     "./bricks/effects/AddTextSnippets.ts",
+    "./bricks/effects/InsertAtCursorEffect.test.ts",
+    "./bricks/effects/InsertAtCursorEffect.ts",
     "./bricks/effects/cancel.ts",
     "./bricks/effects/clipboard.ts",
     "./bricks/effects/confetti.ts",
@@ -70,8 +72,6 @@
     "./bricks/effects/telemetry.ts",
     "./bricks/effects/vue.ts",
     "./bricks/effects/wait.ts",
-    "./bricks/effects/InsertAtCursorEffect.ts",
-    "./bricks/effects/InsertAtCursorEffect.test.ts",
     "./bricks/errors.ts",
     "./bricks/readers/ArrayCompositeReader.ts",
     "./bricks/readers/BlankReader.ts",
@@ -124,8 +124,8 @@
     "./components/Markdown.tsx",
     "./components/MarkdownInline.tsx",
     "./components/ModalLayout.tsx",
-    "./components/StopPropagation.tsx",
     "./components/StopPropagation.test.tsx",
+    "./components/StopPropagation.tsx",
     "./components/Stylesheets.tsx",
     "./components/StylesheetsContext.ts",
     "./components/TooltipIconButton.tsx",
@@ -252,9 +252,9 @@
     "./contrib/automationanywhere/aaTypes.ts",
     "./contrib/automationanywhere/aaUtils.ts",
     "./contrib/automationanywhere/contract.ts",
+    "./contrib/ckeditor/ckeditor.test.ts",
     "./contrib/ckeditor/ckeditorDom.ts",
     "./contrib/ckeditor/ckeditorProtocol.ts",
-    "./contrib/ckeditor/ckeditor.test.ts",
     "./contrib/google/sheets/core/getSheetIdIntegrationOutputKey.ts",
     "./contrib/google/sheets/core/handleGoogleRequestRejection.ts",
     "./contrib/google/sheets/core/schemas.ts",
@@ -269,6 +269,10 @@
     "./contrib/uipath/typeUtils.ts",
     "./contrib/uipath/uipathContract.ts",
     "./contrib/zapier/contract.ts",
+    "./data/service/apiClient.ts",
+    "./data/service/baseService.ts",
+    "./data/service/constants.ts",
+    "./data/service/requestErrorUtils.ts",
     "./development/errorsBadge.ts",
     "./development/hooks/useMessengerLogging.ts",
     "./development/hooks/useRenderCount.ts",
@@ -310,6 +314,7 @@
     "./extensionConsole/pages/workshop/workshopUtils.ts",
     "./extensionConsole/toggleSidebar.tsx",
     "./extensionContext.ts",
+    "./extensionPages/extensionPagePlatform.ts",
     "./globals.d.ts",
     "./hooks/useAsyncExternalStore.ts",
     "./hooks/useAsyncState.ts",
@@ -328,8 +333,8 @@
     "./hooks/useQuickbarShortcut.ts",
     "./hooks/useReduxState.ts",
     "./hooks/useReportError.ts",
-    "./hooks/useScrollLock.ts",
     "./hooks/useScrollLock.test.ts",
+    "./hooks/useScrollLock.ts",
     "./hooks/useSetDocumentTitle.ts",
     "./hooks/useTimeoutState.ts",
     "./hooks/useUndo.ts",
@@ -349,6 +354,7 @@
     "./integrations/sanitizeIntegrationConfig.ts",
     "./integrations/store/integrationsMigrations.ts",
     "./integrations/store/integrationsSelectors.ts",
+    "./integrations/store/integrationsSlice.test.ts",
     "./integrations/store/integrationsSlice.ts",
     "./integrations/util/collectExistingConfiguredDependenciesForMod.ts",
     "./integrations/util/extractIntegrationIdsFromSchema.ts",
@@ -366,7 +372,6 @@
     "./mods/hooks/useMarketplaceUrl.ts",
     "./mv3/SessionStorage.ts",
     "./mv3/api.ts",
-    "./mv3/extensionPagePlatform.ts",
     "./pageEditor/NonScriptablePage.tsx",
     "./pageEditor/baseFormStateTypes.ts",
     "./pageEditor/components/LoginCard.tsx",
@@ -445,10 +450,6 @@
     "./sandbox/messenger/executor.ts",
     "./sandbox/messenger/registration.ts",
     "./sandbox/sandbox.ts",
-    "./data/service/apiClient.ts",
-    "./data/service/baseService.ts",
-    "./data/service/constants.ts",
-    "./data/service/requestErrorUtils.ts",
     "./sidebar/DefaultPanel.tsx",
     "./sidebar/RendererComponent.tsx",
     "./sidebar/SidebarErrorBoundary.tsx",
@@ -461,7 +462,6 @@
     "./sidebar/modLauncher/constants.ts",
     "./sidebar/protocol.tsx",
     "./sidebar/sidePanel.tsx",
-    "./sidebar/sidebarPlatform.ts",
     "./sidebar/sidebarSelectors.ts",
     "./sidebar/sidebarSlice.ts",
     "./sidebar/thunks/addFormPanel.ts",
@@ -499,6 +499,7 @@
     "./store/sessionChanges/sessionChangesTypes.ts",
     "./store/settings/settingsMigrations.ts",
     "./store/settings/settingsSelectors.ts",
+    "./store/settings/settingsSlice.test.ts",
     "./store/settings/settingsSlice.ts",
     "./store/settings/settingsStorage.ts",
     "./store/settings/settingsTypes.ts",
@@ -642,8 +643,6 @@
     "./vendors/reactPerformanceTesting/utils/pushTask.ts",
     "./vendors/reactPerformanceTesting/utils/shouldTrack.ts",
     "./vendors/reactPerformanceTesting/utils/symbols.ts",
-    "./vendors/validateUuid.ts",
-    "./integrations/store/integrationsSlice.test.ts",
-    "./store/settings/settingsSlice.test.ts"
+    "./vendors/validateUuid.ts"
   ]
 }
